rename templates/windows.js -> unsupported-platforms/windows-csharp.js
